refactor walker components
