+++ client/src/components/Walker/WalkerAppointments.tsx
@@ -1,0 +1,3 @@
+export default function WalkerAppointments({ user, userParam }: any) {
+  return <div>walker appts</div>;
+}
